collision work

+++ PixelariaEngine.Core/Time.cs
@@ -19,7 +19,7 @@ public static class Time
 
     public static float AltTimeScale { get; } = 1f;
 
-    public static uint FrameCount { get; private set; }
+    public static ulong FrameCount { get; private set; }
 
     public static int FrameRate { get; private set; }
